refactor: move AppleTVSelects, AppleTVSensors to entities module
