+++ intg-appletv/utils.py
@@ -5,22 +5,8 @@ Utility functions used for Apple TV integration.
 :license: Mozilla Public License Version 2.0, see LICENSE for more details.
 """
 
-from enum import Enum, StrEnum
+from enum import Enum
 from typing import Any, Type
-
-
-class AppleTVSelects(StrEnum):
-    """Apple TV select values."""
-
-    SELECT_APP = "select_app"
-    SELECT_AUDIO_OUTPUT = "select_audio_output"
-
-
-class AppleTVSensors(StrEnum):
-    """Apple TV sensor values."""
-
-    SENSOR_APP = "sensor_app"
-    SENSOR_AUDIO_OUTPUT = "sensor_audio_output"
 
 
 def filter_attributes(attributes, attribute_type: Type[Enum]) -> dict[str, Any]:
